+++ src/renderer/src/components/SettingsPane.tsx
@@ -182,6 +182,8 @@ export function SettingsPane(props: {
   setMdLineNumbers: (v: boolean) => void
   naturalSort: boolean
   setNaturalSort: (v: boolean) => void
+  graph: boolean
+  setGraph: (v: boolean) => void
   onReset: () => void
   locale: Locale
   setLocale: (v: Locale) => void
@@ -343,6 +345,7 @@ export function SettingsPane(props: {
             />
             <CheckRow label={msg.settings.wordWrap} checked={props.wrap} onChange={props.setWrap} />
             <CheckRow label={msg.settings.wordHighlight} checked={props.wordDiff} onChange={props.setWordDiff} />
+            <CheckRow label={msg.settings.commitGraph} checked={props.graph} onChange={props.setGraph} />
             <CheckRow label={msg.settings.markdownOutline} checked={props.mdOutline} onChange={props.setMdOutline} />
             <CheckRow label={msg.settings.markdownLineNumbers} checked={props.mdLineNumbers} onChange={props.setMdLineNumbers} />
             <Segmented